Split standalone concurrentstack examples into separate files
